+++ src/form/dcp.jl
@@ -251,6 +251,14 @@ function add_bus_voltage_setpoint(sol, pm::GenericPowerModel{T}) where T <: Abst
 end
 
 ""
+function add_storage_setpoint(sol, pm::GenericPowerModel{T}) where T <: AbstractDCPForm
+    add_setpoint(sol, pm, "storage", "ps", :ps)
+    add_setpoint_fixed(sol, pm, "storage", "qs")
+    add_setpoint(sol, pm, "storage", "se", :se, conductorless=true)
+end
+
+
+""
 function variable_voltage_on_off(pm::GenericPowerModel{T}; kwargs...) where T <: AbstractDCPForm
     variable_voltage_angle(pm; kwargs...)
 end