+++ CHANGELOG.md
@@ -2,6 +2,7 @@ PowerModels.jl Change Log
 =========================
 
 ### Staged
+- Add support for a generic storage component
 - Added add_setpoint_fixed to improve solution reporting performance
 - Added support for a user provided JuMP model
 - Improved the Matpower data file export function